+++ sdk/net/http/http2/http_static_table.cpp
@@ -67,29 +67,5 @@ size_t http_static_table::size() { return _static_table.size(); }
 
 void http_static_table::load() {}
 
-hpack_static_table hpack_static_table::_instance;
-
-hpack_static_table* hpack_static_table::get_instance() {
-    _instance.load();
-    return &_instance;
-}
-
-hpack_static_table::hpack_static_table() : http_static_table() {}
-
-void hpack_static_table::load() {
-    if (_static_table.empty()) {
-        critical_section_guard guard(_lock);
-        if (_static_table.empty()) {
-            // RFC 7541 Appendix A.  Static Table Definition
-            // if (_static_table.empty()) ...
-            auto lambda = [&](uint32 index, const char* name, const char* value) -> void {
-                _static_table.insert(std::make_pair(name, std::make_pair(value ? value : "", index)));
-                _static_table_index.insert(std::make_pair(index, std::make_pair(name, value ? value : "")));
-            };
-            http_resource::get_instance()->for_each_hpack_static_table(lambda);
-        }
-    }
-}
-
 }  // namespace net
 }  // namespace hotplace