+++ test/quic/sample.cpp
@@ -1,0 +1,63 @@
+/* vim: set tabstop=4 shiftwidth=4 softtabstop=4 expandtab smarttab : */
+/**
+ * @file {file}
+ * @author Soo Han, Kim (princeb612.kr@gmail.com)
+ * @desc
+ * @remarks
+ *          RFC 9204 QPACK: Field Compression for HTTP/3
+ *
+ * Revision History
+ * Date         Name                Description
+ *
+ */
+
+#include <stdio.h>
+
+#include <iostream>
+#include <sdk/net/http/http3/qpack.hpp>
+#include <sdk/sdk.hpp>
+
+using namespace hotplace;
+using namespace hotplace::crypto;
+using namespace hotplace::io;
+using namespace hotplace::net;
+
+test_case _test_case;
+t_shared_instance<logger> _logger;
+
+typedef struct _OPTION {
+    int verbose;
+
+    _OPTION() : verbose(0) {
+        // do nothing
+    }
+} OPTION;
+t_shared_instance<t_cmdline_t<OPTION>> _cmdline;
+
+int main(int argc, char** argv) {
+#ifdef __MINGW32__
+    setvbuf(stdout, 0, _IOLBF, 1 << 20);
+#endif
+
+    _cmdline.make_share(new t_cmdline_t<OPTION>);
+    *_cmdline << t_cmdarg_t<OPTION>("-v", "verbose", [](OPTION& o, char* param) -> void { o.verbose = 1; }).optional();
+    _cmdline->parse(argc, argv);
+
+    const OPTION& option = _cmdline->value();
+
+    logger_builder builder;
+    builder.set(logger_t::logger_stdout, option.verbose).set(logger_t::logger_flush_time, 0).set(logger_t::logger_flush_size, 0);
+    _logger.make_share(builder.build());
+
+    openssl_startup();
+
+    // studying ...
+
+    openssl_cleanup();
+
+    _logger->flush();
+
+    _test_case.report(5);
+    _cmdline->help();
+    return _test_case.result();
+}